requirements.txt
flask==2.0.1
werkzeug==2.0.1
pycryptodome==3.10.1
requests==2.26.0
jinja2==3.0.1
itsdangerous==2.0.1
